+++ src/Symfony/Component/Intl/Tests/Data/Provider/AbstractScriptDataProviderTest.php
@@ -213,7 +213,6 @@ abstract class AbstractScriptDataProviderTest extends AbstractDataProviderTest
         'Zsym',
         'Zxxx',
         'Zyyy',
-        'Zzzz',
     ];
 
     /**