+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -8,7 +8,7 @@ CHANGELOG
  * added `Currencies` in favor of `Intl::getCurrencyBundle()`
  * added `Languages` and `Scripts` in favor of `Intl::getLanguageBundle()`
  * added `Locales` in favor of `Intl::getLocaleBundle()`
- * added `Regions` in favor of `Intl::getRegionBundle()`
+ * added `Countries` in favor of `Intl::getRegionBundle()`
  * added `Timezones`
 
 4.2.0