+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -10,6 +10,8 @@ CHANGELOG
  * added `Locales` in favor of `Intl::getLocaleBundle()`
  * added `Countries` in favor of `Intl::getRegionBundle()`
  * added `Timezones`
+ * made country codes ISO 3166 compliant
+ * excluded script code `Zzzz`
 
 4.2.0
 -----